74 publishing rules (#78)
* Moving all examples to dedicated project

* Ensure common version of Scala in both projects

* Updated DockerFile and pattern example.
Need to test Docker image consistency.
Check that publish does not include examples

* Linter setup

* Solve docker error

* Add Keystone to examples
Add semantic distribution test for Keystone
Move to Maven distributed javaBDD version

* Add Keystone to Docker

* Moving to Java 17

+++ src/main/scala/onera/pmlanalyzer/views/interference/model/formalisation/BDDFactory.scala
@@ -17,7 +17,7 @@
 
 package onera.pmlanalyzer.views.interference.model.formalisation
 
-import net.sf.javabdd.{BDD => JavaBDD, BDDFactory => JavaBDDFactory}
+import com.github.javabdd.{BDD => JavaBDD, BDDFactory => JavaBDDFactory}
 
 import scala.collection.mutable
 
@@ -608,7 +608,7 @@ trait GenBDDFactory[Var] extends BaseBDDFactory[Var, JavaBDD] {
 
   def getPathCount(bdd: JavaBDD, weights: Map[Var, Int] = Map.empty): BigInt = {
     if (weights.isEmpty)
-      BigInt(bdd.pathCount().toLong)
+      bdd.pathCount()
     else {
       val _memo = mutable.HashMap.empty[JavaBDD, BigInt]
